Using long for times

+++ AndroidFITS/app/src/main/java/ie/dit/d13122842/main/History.java
@@ -3,19 +3,19 @@ package ie.dit.d13122842.main;
 public class History {
 
     private static int unitCount = 0;
-    private static int averageTime = 0;
+    private static long averageTime = 0;
 
     public static int getUnitCount() {
         return unitCount;
     }
 
-    public static int getAverageTime() throws Exception {
+    public static long getAverageTime() throws Exception {
         if (unitCount == 0)
             throw new Exception("Attempt to get average time when no units have been timed.");
         return averageTime;
     }
 
-    public static void insert(int value) {
+    public static void insert(long value) {
         averageTime = (unitCount * averageTime + value) / (unitCount + 1);
         unitCount++;
     }
